Cancel all old workflows from first new workflow
Remove specific `${{ github.workflow }}` in the 'concurrency' group, so that any job starting on a branch will cancel all old jobs. Useful for when the CI is busy, so it may take some time for all jobs to start. A `${{ github.workflow }}` probably should be added if we have a documentation-building CI in future, which we might want to trigger on its own while tests still run on a previous commit.