Better default handling for mxg, myg in get_processor_layout()
When `mxg` or `myg` are not given explicitly as arguments to
`processor_rearrange.get_processor_layout()`, try to read them from the
`boutfile`, and only set to hard-coded value of `2` if they are not
present in `boutfile`.

Fixes a bug with `restart.redistribute()` where if the input restart
files had `mxg != 2` or `myg != 2` the output restart files would be
wrong.