Change fill_value from default `Nan` to `0.0`
When resizing from a fine grid to a coarser grid, the location of the boundary guard cells will be to the left/right of the left/right guard cell of the finer grid. This is extrapolation which by default the RegularGridInterpolator method sets to `nan`. BOUT++ might struggle to interpret this at run time so I have changed this to `0.0`